requirements.txt
# Core API
fastapi==0.110.0
uvicorn==0.29.0
pydantic==2.6.4
httpx==0.27.0

# LLM / Agents
openai==1.14.0
langgraph==0.0.38
langchain==0.1.13

# HuggingFace ecosystem
transformers==4.39.0
peft==0.10.0
accelerate==0.27.2
datasets==2.18.0
sentence-transformers==2.6.1
bitsandbytes==0.43.0

# Vector store
faiss-cpu==1.8.0

# Distributed / async
celery==5.3.6
redis==5.0.3
ray[train]==2.10.0

# ML / data science
torch==2.2.1
torchvision==0.17.1
numpy==1.26.4
scipy==1.12.0
scikit-learn==1.4.1
pandas==2.2.1
matplotlib==3.8.3
umap-learn==0.5.6

# MLOps
mlflow==2.11.1
ragas==0.1.7

# AWS
boto3==1.34.0
sagemaker==2.214.0

# Data ingestion
pyspark==3.5.1
warcio==1.7.4
langdetect==1.0.9
datasketch==1.6.4

# Diffusion / multimodal
diffusers==0.27.2
pillow==10.2.0

# MCP server
mcp>=1.0.0

# Testing / linting
pytest==8.1.1
pytest-asyncio==0.23.6
pytest-cov==5.0.0
ruff==0.3.4
black==24.3.0
